[Day 13.3] 게시물 삭제시 권한에 따른 제한"

+++ demo/src/main/java/com/study/demo/controller/BoardController.java
@@ -98,9 +98,11 @@ public class BoardController {
         // 로그인 정보 확인
         Claims token = jwtService.checkAuthorizationHeader(request);
         if(token == null) return null;
-        System.out.println(token);
+        String writerId = boardService.getBoard(id).getUser().getUserid();
+        if(!writerId.equals(token.get("userId").toString())){
+            return null;
+        }
         boardService.delete(id);
-
         return "success";
     }
